updated the module to work on server and app
