+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "apogeejs-module-sqlite",
-    "version": "0.0.1",
+    "version": "0.0.2",
     "description": "ApogeeJS node module for SQLite access",
     "main": "./src/SqliteComponentModule.js",
     "dependencies": {
